third_party/blink/web_tests/platform/linux/http/tests/persistent-origin-trials/critical-origin-trial.https-expected.txt

This is a testharness.js-based test.
FAIL PersistentOriginTrial using Critical header assert_equals: expected "FrobulatePersistent" but got ""
Harness: the test ran to completion.
